backend/requirements.txt
fastapi
uvicorn
supabase
stripe
python-dotenv
pydantic
redis>=5.0.0
